ARTHA_AI/requirements.txt
streamlit>=1.32.0
streamlit-autorefresh>=0.0.1
yfinance>=0.2.38
pandas>=2.0.0
numpy>=1.26.0
plotly>=5.20.0
moviepy>=1.0.3
Pillow>=10.0.0
matplotlib>=3.8.0
requests>=2.31.0
